updated Notepad
