+++ client/src/types/NotepadTypes.ts
@@ -1,0 +1,8 @@
+export interface Note {
+    note_id: string,
+    title: string,
+    content: string,
+    color: string,
+    created_date: string,
+    created_time: string,
+  }
